requirements.txt
httpx>=0.27.0
tenacity>=8.5.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pandas>=2.0.0
fake-useragent>=1.5.0
playwright>=1.48.0
click>=8.0.0
python-dotenv>=1.0.0